Move router source and classes into world library too.
