+++ src/avitab/apps/RouteApp.h
@@ -22,7 +22,7 @@
 #include <vector>
 #include <set>
 #include "App.h"
-#include "src/libxdata/router/Route.h"
+#include "src/world/router/Route.h"
 #include "src/gui_toolkit/widgets/TabGroup.h"
 #include "src/gui_toolkit/widgets/Page.h"
 #include "src/gui_toolkit/widgets/TextArea.h"
@@ -52,7 +52,7 @@ private:
     world::AirwayLevel airwayLevel = world::AirwayLevel::UPPER;
     std::shared_ptr<world::Airport> departureAirport, arrivalAirport;
     std::shared_ptr<world::Fix> departureFix, arrivalFix;
-    std::shared_ptr<xdata::Route> route;
+    std::shared_ptr<world::Route> route;
 
     void showDeparturePage();
     void onDepartureEntered(const std::string &departure);
