+++ src/libxdata/CMakeLists.txt
@@ -6,4 +6,3 @@ target_link_libraries(xdata PUBLIC world)
 
 include(${CMAKE_CURRENT_LIST_DIR}/parsers/CMakeLists.txt)
 include(${CMAKE_CURRENT_LIST_DIR}/loaders/CMakeLists.txt)
-include(${CMAKE_CURRENT_LIST_DIR}/router/CMakeLists.txt)
