--- src/libxdata/router/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(xdata PRIVATE
-    ${CMAKE_CURRENT_LIST_DIR}/RouteFinder.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/Route.cpp
-)
