+++ src/world/router/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(world PRIVATE
+    ${CMAKE_CURRENT_LIST_DIR}/RouteFinder.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/Route.cpp
+)
